# spec/openapi.yaml
openapi: "3.0.2"
info:
  title: gohost
  description: Go Host!
  version: 0.0.1
paths:
  /register:
    get:
      summary: "ユーザー登録に必要な情報"
      tags:
        - "auth"
      responses:
        200:
          $ref: "#/components/responses/getRegisterResponse"
    post:
      summary: "新規ユーザー登録"
      tags:
        - "auth"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
      responses:
        200:
          description: "登録成功"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
  /login:
    post:
      summary: "ログイン"
      tags:
        - "auth"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        200:
          description: "ログイン成功"
          headers:
            Set-Cookie:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
  /logout:
    post:
      summary: "ログアウト"
      tags:
        - "auth"
      responses:
        200:
          description: "ログアウト成功"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
  /spots:
    get:
      summary: "肝試しスポットの一覧取得"
      tags:
        - "match"
      parameters:
        - in: query
          required: true
          name: date
          schema:
            type: string
            format: date
        - in: query
          required: false
          name: limit
          schema:
            type: integer
            default: 3
      responses:
        200:
          description: "ok"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Spot"
  /spots/{spot_id}:
    parameters:
      - in: path
        required: true
        name: spot_id
        schema:
          type: string
    get:
      summary: "肝試しスポット"
      tags:
        - "match"
      responses:
        200:
          description: "ok"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Spot"
  /spots/{spot_id}/entry:
    parameters:
      - in: path
        required: true
        name: spot_id
        schema:
          type: string
    post:
      summary: "どの肝試しスポットにするかを決定"
      tags:
        - "match"
      requestBody:
        description: エントリー情報
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                date:
                  type: string
                  format: date
      responses:
        200:
          description: "ok"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
  /plan:
    get:
      summary: "肝試しの確定日程の情報"
      tags:
        - "plan"
      responses:
        200:
          $ref: "#/components/responses/getPlanResponse"
  /plan/cancel:
    post:
      summary: "肝試しの日程キャンセル"
      tags:
        - "plan"
      responses:
        200:
          description: "ok"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
  /plan/finish:
    get:
      summary: "肝試しを終了する(メンバー取得)"
      tags:
        - "plan"
      responses:
        200:
          $ref: "#/components/responses/getFinishPlanResponse"
    post:
      summary: "肝試しを終了する(メンバー評価)"
      tags:
        - "plan"
      requestBody:
        $ref: "#/components/requestBodies/postFinishPlanRequestBody"
      responses:
        200:
          description: "ok"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        user_name:
          type: string
        email:
          type: string
        password:
          type: string
        university_id:
          type: string
        birth_date:
          type: string
          format: date
        year:
          type: integer
          description: "学年"
        gender_id:
          type: string
        icon_url:
          type: string
          description: "アイコンのURLなど"
        instagram_id:
          type: string
          description: "インスタグラムのID"
    Spot:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        image_url:
          type: string
          description: "心霊スポットのURL"
        address:
          type: string
          description: "住所"
        prefecture:
          type: string
          description: "都道府県"
    Message:
      type: object
      properties:
        message:
          type: string
    Gender:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
    University:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        meeting_station:
          type: string
  requestBodies:
    postFinishPlanRequestBody:
      description: ユーザーへのgood
      required: true
      content:
        application/json:
          schema:
            type: array
            items:
              type: object
              properties:
                user_id:
                  type: string
                like:
                  type: boolean
  responses:
    getRegisterResponse:
      description: "ok"
      content:
        application/json:
          schema:
            type: object
            properties:
              universities:
                type: array
                items:
                  $ref: "#/components/schemas/University"
              genders:
                type: array
                items:
                  $ref: "#/components/schemas/Gender"
    getPlanResponse:
      description: "ok"
      content:
        application/json:
          schema:
            type: object
            properties:
              meeting_station:
                type: string
                description: "集合駅(一覧画面では返さない)"
              meeting_hour:
                type: string
                format: "date-time"
                description: "集合時間(一覧画面では返さない)"
              spot:
                $ref: "#/components/schemas/Spot"
              users:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    getFinishPlanResponse:
      description: "ok"
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: "#/components/schemas/User"
security:
  - cookieAuth: []